Fix: Rounding on `getBlock` on negative coords
If the getBlock method was being accessed with the Pos3D class it was using different coordinates than the method with the BlockPosHelper as they both rounded differently.
The Pos3D truncated just the different values while the BlockPosHelper floored the different values (which is the normal Minecraft way).

This means that the Pos3D way was incorrect.